requirements.txt
flask==3.0.0
openai==1.3.5
google-auth-oauthlib==1.0.0
google-api-python-client==2.95.0
psutil==5.9.5
python-dotenv==1.0.0
